refactor: basicComment
