+++ packages/client/src/pages/forum/components/CommentWithReply/index.tsx
@@ -1,10 +1,4 @@
-import {
-  DislikeFilled,
-  DislikeOutlined,
-  LikeFilled,
-  LikeOutlined,
-} from '@ant-design/icons';
-import { Avatar, Comment as Commentary, Tooltip } from 'antd';
+
 import { useState } from 'react';
 import createCn from '../../../../utils/create-cn';
 import { TBasicComment, TCommentWithReply } from '../../types';
@@ -19,32 +13,12 @@ type Props = {
 const cn = createCn('comment');
 
 function CommentWithReply({ comment }: Props) {
-  const {
-    author,
-    avatar = 'https://joeschmoe.io/api/v1/random',
-    content,
-    datetime,
-    subComments,
-  } = comment;
-  const [likes, setLikes] = useState(0);
-  const [dislikes, setDislikes] = useState(0);
-  const [action, setAction] = useState<'liked' | 'disliked' | null>(null);
+  const { subComments } = comment;
+
   const [isOpenEditor, setIsOpenEditor] = useState<boolean>(false);
   const [newCommentvalue, setNewCommentvalue] = useState('');
   const [submitting, setSubmitting] = useState(false);
   const [subCommentsArr, setSubCommentsArr] = useState<TBasicComment[]>(subComments);
-
-  const like = () => {
-    setLikes(1);
-    setDislikes(0);
-    setAction('liked');
-  };
-
-  const dislike = () => {
-    setLikes(0);
-    setDislikes(1);
-    setAction('disliked');
-  };
 
   const toogleEditor = () => setIsOpenEditor(!isOpenEditor);
 
@@ -74,19 +48,7 @@ function CommentWithReply({ comment }: Props) {
     }, 1000);
   };
 
-  const actions = [
-    <Tooltip key="comment-like" title="Нравится">
-      <button className={cn('button-action')} type="button" onClick={like}>
-        {action === 'liked' ? <LikeFilled /> : <LikeOutlined />}
-        <span className={cn('comment-action')}>{likes}</span>
-      </button>
-    </Tooltip>,
-    <Tooltip key="comment-dislike" title="Не нравится">
-      <button className={cn('button-action')} type="button" onClick={dislike}>
-        {action === 'disliked' ? <DislikeFilled /> : <DislikeOutlined />}
-        <span className={cn('comment-action')}>{dislikes}</span>
-      </button>
-    </Tooltip>,
+  const additionalActions = [
     <button
       key="comment-reply"
       className={cn('button-reply')}
@@ -98,13 +60,7 @@ function CommentWithReply({ comment }: Props) {
 
   return (
     <div className={cn()}>
-      <Commentary
-        actions={actions}
-        author={<p>{author}</p>}
-        avatar={<Avatar src={avatar} alt={`Аватар ${author}`} />}
-        content={content}
-        datetime={datetime}
-      />
+      <BasicComment comment={comment} additionalActions={additionalActions}/>
       <div className={cn('sub-comments')}>
         {isOpenEditor && (
           <Editor
@@ -114,7 +70,7 @@ function CommentWithReply({ comment }: Props) {
             value={newCommentvalue}
           />
         )}
-        {subCommentsArr.map((subComment) => <BasicComment key={subComment.id} comment={subComment} />)}
+        {subCommentsArr.map((subComment) => <BasicComment key={subComment.id} comment={subComment} additionalActions={[]} />)}
       </div>
     </div>
   );
